Fixed type declarations

+++ core/apriltag_demo.c
@@ -109,7 +109,7 @@ int main(int argc, char *argv[])
         if (maxiters > 1)
             printf("iter %d / %d\n", iter + 1, maxiters);
 
-        for (int input = 0; input < zarray_size(inputs); input++) {
+        for (size_t input = 0; input < zarray_size(inputs); input++) {
 
             int hamm_hist[hamm_hist_max];
             memset(hamm_hist, 0, sizeof(hamm_hist));
@@ -135,14 +135,14 @@ int main(int argc, char *argv[])
 
             total_detections += zarray_size(detections);
 
-            for (int i = 0; i < zarray_size(detections); i++) {
+            for (size_t i = 0; i < zarray_size(detections); i++) {
                 apriltag_detection_t *det;
                 zarray_get(detections, i, &det);
 
                 if (benchmark) {
                     printf(" %d", det->id);
                 } else if (!quiet) {
-                    printf("detection %3d: id (%2dx%2d)-%-4d, hamming %d, goodness %8.3f, margin %8.3f\n",
+                    printf("detection %3ld: id (%2dx%2d)-%-4d, hamming %d, goodness %8.3f, margin %8.3f\n",
                            i, det->family->d*det->family->d, det->family->h, det->id, det->hamming, det->goodness, det->decision_margin);
                 }
 
@@ -179,7 +179,7 @@ int main(int argc, char *argv[])
     }
 
     if (benchmark) {
-        fprintf(stderr, "%d detections over %d images in %.3f ms (%.3f ms per frame)\n",
+        fprintf(stderr, "%d detections over %ld images in %.3f ms (%.3f ms per frame)\n",
                 total_detections, zarray_size(inputs),
                 (total_time*1e-3), (total_time*1e-3)/zarray_size(inputs));
     }
